MORE

+++ lib/src/fd.rs
@@ -5,7 +5,7 @@ pub struct FileDescriptorId(pub usize);
 
 bitflags! {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    pub struct FileDescriptorFlags: u32 {
+    pub struct FileDescriptorFlags: usize {
         const O_RDONLY = 0x1;
         const O_WRONLY = 0x2;
         const O_RDWR = Self::O_RDONLY.bits() | Self::O_WRONLY.bits();
